+++ internal/resigner/listener_test.go
@@ -141,7 +141,12 @@ func TestRun(t *testing.T) {
 
 	queue <- &securityv1.Resource{
 		Config: &securityv1.Resource_Api{
-			Api: structpb.NewStructValue(apiStruct),
+			Api: &apiv1.Api{
+				Signature: &utilsv1.Signature{
+					KeyId: "example",
+					Data:  sig,
+				},
+			},
 		},
 		GitRef: &securityv1.Resource_BranchName{
 			BranchName: "main",