+++ internal/signature/reconciler/reconciler_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/jonboulle/clockwork"
 	"github.com/superblocksteam/agent/pkg/utils"
 	pbapi "github.com/superblocksteam/agent/types/gen/go/api/v1"
+	pbcommon "github.com/superblocksteam/agent/types/gen/go/common/v1"
 	pbsecurity "github.com/superblocksteam/agent/types/gen/go/security/v1"
 	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/testing/protocmp"
@@ -91,15 +92,12 @@ func validArgs(t *testing.T) *args {
 
 		switch i % typesOfResources {
 		case 0:
-			api, err := structpb.NewStruct(map[string]any{
-				"metadata": map[string]any{
-					"id": fmt.Sprintf("%d", i),
-				},
-			})
-			require.NoError(t, err)
-
 			res.Config = &pbsecurity.Resource_Api{
-				Api: structpb.NewStructValue(api),
+				Api: &pbapi.Api{
+					Metadata: &pbcommon.Metadata{
+						Id: fmt.Sprintf("%d", i),
+					},
+				},
 			}
 		case 1:
 			literal, err := structpb.NewStruct(map[string]any{
@@ -278,18 +276,15 @@ func TestNoResourcesToSign(t *testing.T) {
 func TestOnlyApisToSign(t *testing.T) {
 	t.Parallel()
 
-	api, err := structpb.NewStruct(map[string]any{
-		"metadata": map[string]any{
-			"id": "0",
-		},
-	})
-	require.NoError(t, err)
-
 	args := validArgs(t)
 	args.server.ClaimBatchToSignSet([]*pbsecurity.Resource{
 		{
 			Config: &pbsecurity.Resource_Api{
-				Api: structpb.NewStructValue(api),
+				Api: &pbapi.Api{
+					Metadata: &pbcommon.Metadata{
+						Id: "0",
+					},
+				},
 			},
 			GitRef: &pbsecurity.Resource_CommitId{
 				CommitId: "0",
@@ -641,8 +636,10 @@ func TestPatchFromApiLiteralMissingMetadata(t *testing.T) {
 	require.NoError(t, err)
 
 	resource := &pbsecurity.Resource{
-		Config: &pbsecurity.Resource_Api{
-			Api: structpb.NewStructValue(apiNoMetadata),
+		Config: &pbsecurity.Resource_ApiLiteral_{
+			ApiLiteral: &pbsecurity.Resource_ApiLiteral{
+				Data: structpb.NewStructValue(apiNoMetadata),
+			},
 		},
 	}
 	err = args.signer.SignAndUpdateResource(resource)
@@ -664,8 +661,10 @@ func TestPatchFromApiLiteralMissingSignature(t *testing.T) {
 	require.NoError(t, err)
 
 	resource := &pbsecurity.Resource{
-		Config: &pbsecurity.Resource_Api{
-			Api: structpb.NewStructValue(apiNoSig),
+		Config: &pbsecurity.Resource_ApiLiteral_{
+			ApiLiteral: &pbsecurity.Resource_ApiLiteral{
+				Data: structpb.NewStructValue(apiNoSig),
+			},
 		},
 	}