+++ internal/signature/reconciler/signer/fakes_test.go
@@ -66,11 +66,7 @@ func (f *fakeSignerManager) SigningKeyID() string {
 func getSignature(res *pbsecurity.Resource) []byte {
 	switch t := res.Config.(type) {
 	case *pbsecurity.Resource_Api:
-		sig, err := signature.StructpbToSignatureProto(t.Api.GetStructValue().GetFields()["signature"].GetStructValue())
-		if err != nil {
-			return []byte("invalid-type-no-signature-" + uuid.NewString())
-		}
-		return sig.GetData()
+		return t.Api.GetSignature().GetData()
 	case *pbsecurity.Resource_ApiLiteral_:
 		sigStruct := t.ApiLiteral.GetData().GetStructValue().GetFields()["signature"].GetStructValue()
 		sig, err := signature.StructpbToSignatureProto(sigStruct)
@@ -90,13 +86,7 @@ func getSignature(res *pbsecurity.Resource) []byte {
 func setSignature(res *pbsecurity.Resource, sig *pbutils.Signature) {
 	switch t := res.Config.(type) {
 	case *pbsecurity.Resource_Api:
-		if sig == nil {
-			t.Api.GetStructValue().Fields["signature"] = structpb.NewNullValue()
-			break
-		}
-
-		sigStruct := signature.SignatureProtoToStructpb(sig)
-		t.Api.GetStructValue().Fields["signature"] = structpb.NewStructValue(sigStruct)
+		t.Api.Signature = sig
 	case *pbsecurity.Resource_ApiLiteral_:
 		if sig == nil {
 			t.ApiLiteral.Data.GetStructValue().Fields["signature"] = structpb.NewNullValue()