+++ internal/signature/reconciler/signer/signer_test.go
@@ -5,6 +5,8 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
+	pbapi "github.com/superblocksteam/agent/types/gen/go/api/v1"
+	pbcommon "github.com/superblocksteam/agent/types/gen/go/common/v1"
 	pbsecurity "github.com/superblocksteam/agent/types/gen/go/security/v1"
 	pbutils "github.com/superblocksteam/agent/types/gen/go/utils/v1"
 	"go.uber.org/zap/zaptest"
@@ -29,13 +31,6 @@ func validArgs(t *testing.T) *args {
 }
 
 func validParamResourceTypes(t *testing.T) []*paramResourceType {
-	api, err := structpb.NewStruct(map[string]any{
-		"metadata": map[string]any{
-			"id": "0",
-		},
-	})
-	require.NoError(t, err)
-
 	literal, err := structpb.NewStruct(map[string]any{
 		"metadata": map[string]any{
 			"id": "0",
@@ -48,7 +43,11 @@ func validParamResourceTypes(t *testing.T) []*paramResourceType {
 			name: "api",
 			res: &pbsecurity.Resource{
 				Config: &pbsecurity.Resource_Api{
-					Api: structpb.NewStructValue(api),
+					Api: &pbapi.Api{
+						Metadata: &pbcommon.Metadata{
+							Id: "0",
+						},
+					},
 				},
 				GitRef: &pbsecurity.Resource_CommitId{
 					CommitId: "0",